feat: refactor register and update info
