--- helpers/parameter-validator.js
@@ -1,8 +1,0 @@
-
-class ParameterValidator {
-
-}
-
-exports = module.exports = {
-  ParameterValidator
-}
